+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix copy of old material property animation when it is a color property.
 * Fix optimizer always injecting the animated property disambiguation code on merged material blobs that have at least one material that comes from the combined mesh path.
 * Don't generate default properties which have names that are invalid ShaderLab property names. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/81)
+* Fix always disabled renderer components not getting deleted before mesh merging happens.
 
 ## v3.4.1
 ### Changes